+++ socket.go
@@ -1,0 +1,25 @@
+package manager
+
+import (
+	"syscall"
+
+	"golang.org/x/sys/unix"
+)
+
+// attachSocket - Attaches the probe to the provided socket
+func (p *Probe) attachSocket() error {
+	return sockAttach(p.SocketFD, p.program.FD())
+}
+
+// detachSocket - Detaches the probe from its socket
+func (p *Probe) detachSocket() error {
+	return sockDetach(p.SocketFD, p.program.FD())
+}
+
+func sockAttach(sockFd int, progFd int) error {
+	return syscall.SetsockoptInt(sockFd, syscall.SOL_SOCKET, unix.SO_ATTACH_BPF, progFd)
+}
+
+func sockDetach(sockFd int, progFd int) error {
+	return syscall.SetsockoptInt(sockFd, syscall.SOL_SOCKET, unix.SO_DETACH_BPF, progFd)
+}
