+++ syscalls.go
@@ -1,15 +1,2 @@
 package manager
 
-import (
-	"syscall"
-
-	"golang.org/x/sys/unix"
-)
-
-func sockAttach(sockFd int, progFd int) error {
-	return syscall.SetsockoptInt(sockFd, syscall.SOL_SOCKET, unix.SO_ATTACH_BPF, progFd)
-}
-
-func sockDetach(sockFd int, progFd int) error {
-	return syscall.SetsockoptInt(sockFd, syscall.SOL_SOCKET, unix.SO_DETACH_BPF, progFd)
-}
